+++ include/CLUEstering/data_structures/detail/PointsConversion.hpp
@@ -101,7 +101,7 @@ namespace clue {
     if (h_points.view().has_tags()) {
       using PType = std::remove_cvref_t<decltype(d_points)>;
       auto& tbuf = internal::points_interface<PType>::tags_buffer(d_points);
-      tbuf = make_device_buffer<std::size_t[]>(queue, h_points.size());
+      tbuf = make_device_buffer<std::uint32_t[]>(queue, h_points.size());
       alpaka::memcpy(queue,
                      make_device_view(alpaka::getDev(queue), tbuf->data(), h_points.size()),
                      make_host_view(h_points.view().m_tags, h_points.size()));
@@ -149,7 +149,7 @@ namespace clue {
     if (h_points.view().has_tags()) {
       using PType = std::remove_cvref_t<decltype(d_points)>;
       auto& tbuf = internal::points_interface<PType>::tags_buffer(d_points);
-      tbuf = make_device_buffer<std::size_t[]>(queue, h_points.size());
+      tbuf = make_device_buffer<std::uint32_t[]>(queue, h_points.size());
       alpaka::memcpy(queue,
                      make_device_view(alpaka::getDev(queue), tbuf->data(), h_points.size()),
                      make_host_view(h_points.view().m_tags, h_points.size()));
